resolves type warning for format string

+++ NSDate+Helper.m
@@ -70,7 +70,7 @@
 			text = @"Yesterday";
 			break;
 		default:
-			text = [NSString stringWithFormat:@"%d days ago", daysAgo];
+			text = [NSString stringWithFormat:@"%ld days ago", (NSInteger)daysAgo];
 	}
 	return text;
 }
